Fix infinispan tests. Remove mixed-jdbc as it is deprecated.

+++ runFailedTests.sh
@@ -1,8 +1,13 @@
-grep Exception tests/basic/target/surefire-reports/*.txt | awk -F ':' '{print $1}' | sort|uniq | while read dir ; do 
+rm -f failed.sh
+
+grep -E 'FAILURE|Exception' tests/basic/target/surefire-reports/*.txt | awk -F ':' '{print $1}' | sort|uniq | while read dir ; do 
     pkg=`echo $dir|sed 's/tests\/basic\/target\/surefire-reports\///g;s/\.txt//g'`; 
     echo "mvn test -P chrome,basic,standalone -Dtest=$pkg" >> failed.sh;
 done
 
-rm -f tests/basic/target/surefire-reports/*.txt
+cat failed.sh
+if [ "$1" = "-r" ] ; then
+    rm -f tests/basic/target/surefire-reports/*.txt
+    source failed.sh
+fi
 
-source failed.sh
